feat: Integrate Twiddle algorithm inside main loop.
A OPTIMIZATION bool variable is introduced in main to activate
the execution of Twiddle algorithm. If it's set to true for each
iteration the Twiddle Run method is called. If Twiddle tries to change
the PID parameters the simulation is restarted with a restart message.
Otherwise the PID control input is applied.